Final push?

+++ MakersMassiveJuiceIt/Assets/!Scripts/Managers/EventManager.cs
@@ -5,14 +5,14 @@ public static class EventManager
 {
     public static Action OnScreenShake;
 
-    public static Action<int> OnIncreaseScore;
+    public static Action<float> OnIncreaseScore;
 
     public static void TriggerScreenShake()
     {
         OnScreenShake?.Invoke();
     }
 
-    public static void TriggerIncreaseScore(int score)
+    public static void TriggerIncreaseScore(float score)
     {
         OnIncreaseScore?.Invoke(score);
     }
